+++ .gitignore
@@ -397,3 +397,4 @@ FodyWeavers.xsd
 # JetBrains Rider
 *.sln.iml
 2023/AdventOfCode2023/.idea/.idea.AdventOfCode2023/.idea/workspace.xml
+2023/AdventOfCode2023/.idea/.idea.AdventOfCode2023/.idea/workspace.xml
