+++ 2023/AdventOfCode2023/AdventOfCode2023/Input/Day04/MockDay04.in
