--- 2023/AdventOfCode2023/AdventOfCode2023/Models/Day03/EngineNumber.cs
@@ -1,10 +1,0 @@
-﻿namespace AdventOfCode2023_1.Models.Day03;
-
-public class EngineNumber(int rowIndex, int columnIndex, int numberLength, int number)
-{
-    public int RowIndex { get; } = rowIndex;
-    public int ColumnIndex { get; } = columnIndex;
-    public int NumberLength { get; set; } = numberLength;
-    public int Number { get; set; } = number;
-    public bool IsPartNumber { get; set; }
-}
