--- 2023/AdventOfCode2023/AdventOfCode2023/PartsToRun.cs
@@ -1,8 +1,0 @@
-namespace AdventOfCode2023_1;
-
-public enum PartsToRun
-{
-    Part1,
-    Part2,
-    Both
-}
