+++ 2023/AdventOfCode2023/AdventOfCode2023/Program.cs
@@ -2,8 +2,7 @@
 
 using AdventOfCode2023_1;
 
-// new Day01().Run(1,"Trebuchet?!");
-// new Day02().Run(2,"Cube Conundrum");
-// new Day03().Run(3, "Gear Ratios");
-// new Day04().Run(4, "Scratchcards");
-new Day05().Run(5, "If You Give A Seed A Fertilizer", PartsToRun.Part1);
+// Day01.Run();
+// Day02.Run();
+// Day03.Run();
+Day04.Run();
