+++ 2023/AdventOfCode2023/AdventOfCode2023/Shared/Constants.cs
@@ -9,5 +9,4 @@ public static class Constants
     public const string Colon = ":";
     public const string SemiColon = ";";
     public const string Pipe = "|";
-    public const string LineReturn = "\r";
 }
