+++ README.md
@@ -2,15 +2,3 @@
 Collection of Advent Of Code years
 
 <a href="https://adventofcode.com/2023" target="_blank">Advent of code 2023</a>
-
-## Answers
-| Day   | Part | Answer   |
-|-------|------|----------|
-| 01    | 1    | 55971    |
-|     | 2    | 54719    |
-| 02    | 1    | 2447     |
-|     | 2    | 56322    |
-| 03    | 1    | 553079   |
-|     | 2    | 84363105 |
-| 04    | 1    | 24848    |
-|     | 2    | 7258152  |
